📚 docs: add swagger
