+++ src/infra/app.controller.ts
@@ -1,0 +1,13 @@
+import { Controller, Get } from '@nestjs/common'
+// biome-ignore lint/style/useImportType: <explanation>
+import { AppService } from './app.service'
+
+@Controller()
+export class AppController {
+  constructor(private readonly appService: AppService) {}
+
+  @Get()
+  getHello(): string {
+    return this.appService.getHello()
+  }
+}
